update how event handling is done (discordjs docs)
