+++ .env.example
@@ -50,6 +50,8 @@ DEFAULT_TTS_SPEAKER="rocket"
 TOKEN_RESET_TIME=3600000
 # Amount of maximum completion tokens (doesn't include prompt tokens) that can be used in the time above (applies to all users). ie 2000
 TOKEN_NUM=2000
+# Number of messages to keep in conversation history. (includes user and bot messages)
+MSG_LIMIT=10
 
 # Personalities
 # Add as many personalities as you want.
